Moves util function declarations
